refactor(netif): decompose the interfaces god-file by role (WS-D)
Split internal/netif/interfaces.go (604 LOC) into role-based files in the
same package — pure move: interface inventory/management stays in
interfaces.go; auto-detection, candidate selection, and link-status probing
move out.

Both files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
